+++ packages/bottom-tabs/src/views/BottomTabView.tsx
@@ -159,10 +159,15 @@ export function BottomTabView(props: Props) {
         tabBarPosition === 'left'
           ? styles.left
           : tabBarPosition === 'right'
-          ? styles.right
-          : null
+            ? styles.right
+            : null
       }
     >
+      {tabBarPosition === 'top' ? (
+        <BottomTabBarHeightCallbackContext.Provider value={setTabBarHeight}>
+          {renderTabBar()}
+        </BottomTabBarHeightCallbackContext.Provider>
+      ) : null}
       <MaybeScreenContainer
         enabled={detachInactiveScreens}
         hasTwoStates={hasTwoStates}
@@ -214,16 +219,16 @@ export function BottomTabView(props: Props) {
           const activityState = isFocused
             ? STATE_ON_TOP // the screen is on top after the transition
             : animationEnabled // is animation is not enabled, immediately move to inactive state
-            ? tabAnims[route.key].interpolate({
-                inputRange: [0, 1 - EPSILON, 1],
-                outputRange: [
-                  STATE_TRANSITIONING_OR_BELOW_TOP, // screen visible during transition
-                  STATE_TRANSITIONING_OR_BELOW_TOP,
-                  STATE_INACTIVE, // the screen is detached after transition
-                ],
-                extrapolate: 'extend',
-              })
-            : STATE_INACTIVE;
+              ? tabAnims[route.key].interpolate({
+                  inputRange: [0, 1 - EPSILON, 1],
+                  outputRange: [
+                    STATE_TRANSITIONING_OR_BELOW_TOP, // screen visible during transition
+                    STATE_TRANSITIONING_OR_BELOW_TOP,
+                    STATE_INACTIVE, // the screen is detached after transition
+                  ],
+                  extrapolate: 'extend',
+                })
+              : STATE_INACTIVE;
 
           return (
             <MaybeScreen
@@ -259,9 +264,11 @@ export function BottomTabView(props: Props) {
           );
         })}
       </MaybeScreenContainer>
-      <BottomTabBarHeightCallbackContext.Provider value={setTabBarHeight}>
-        {renderTabBar()}
-      </BottomTabBarHeightCallbackContext.Provider>
+      {tabBarPosition !== 'top' ? (
+        <BottomTabBarHeightCallbackContext.Provider value={setTabBarHeight}>
+          {renderTabBar()}
+        </BottomTabBarHeightCallbackContext.Provider>
+      ) : null}
     </SafeAreaProviderCompat>
   );
 }