+++ packages/devtools/src/parseHermesStack.tsx
@@ -70,18 +70,18 @@ function parseLine(line: string): HermesStackEntry | undefined {
         asFrame[2] === 'native'
           ? { type: 'NATIVE' }
           : asFrame[3] === 'address at '
-          ? {
-              type: 'BYTECODE',
-              sourceUrl: asFrame[4],
-              line1Based: Number.parseInt(asFrame[5], 10),
-              virtualOffset0Based: Number.parseInt(asFrame[6], 10),
-            }
-          : {
-              type: 'SOURCE',
-              sourceUrl: asFrame[4],
-              line1Based: Number.parseInt(asFrame[5], 10),
-              column1Based: Number.parseInt(asFrame[6], 10),
-            },
+            ? {
+                type: 'BYTECODE',
+                sourceUrl: asFrame[4],
+                line1Based: Number.parseInt(asFrame[5], 10),
+                virtualOffset0Based: Number.parseInt(asFrame[6], 10),
+              }
+            : {
+                type: 'SOURCE',
+                sourceUrl: asFrame[4],
+                line1Based: Number.parseInt(asFrame[5], 10),
+                column1Based: Number.parseInt(asFrame[6], 10),
+              },
     };
   }
   const asSkipped = line.match(RE_SKIPPED);