+++ packages/react-native-tab-view/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.0.0-alpha.3](https://github.com/react-navigation/react-navigation/compare/react-native-tab-view@4.0.0-alpha.2...react-native-tab-view@4.0.0-alpha.3) (2024-01-17)
+
+**Note:** Version bump only for package react-native-tab-view
+
 # [4.0.0-alpha.2](https://github.com/react-navigation/react-navigation/compare/react-native-tab-view@4.0.0-alpha.1...react-native-tab-view@4.0.0-alpha.2) (2023-11-17)
 
 **Note:** Version bump only for package react-native-tab-view